requirements.txt
python-binance
TA-Lib
numpy
backtrader
flask
websocket_client
gunicorn